+++ apps/docs/src/examples/rapier/rigid-body/Particle.svelte
