+++ apps/docs/src/examples/rapier/rigid-body/Scene.svelte
@@ -3,72 +3,12 @@
 		DirectionalLight,
 		Object3DInstance,
 		OrbitControls,
-		PerspectiveCamera,
-		useFrame,
-		type Position,
-		type Rotation
+		PerspectiveCamera
 	} from '@threlte/core'
-	import { HTML } from '@threlte/extras'
-	import { Debug, RigidBody } from '@threlte/rapier'
-	import { Euler } from 'three'
-	import { Vector3 } from 'three'
+	import { Debug } from '@threlte/rapier'
 	import { GridHelper } from 'three'
-	import { DEG2RAD } from 'three/src/math/MathUtils'
+	import Emitter from './Emitter.svelte'
 	import Ground from './Ground.svelte'
-	import TestBody from './TestBody.svelte'
-
-	const getId = () => {
-		return Math.random().toString(16).slice(2)
-	}
-
-	const getRandomPosition = () => {
-		return new Vector3(Math.random() * 1, Math.random() * 1 + 10, Math.random() * 1)
-	}
-
-	const getRandomRotation = () => {
-		return new Euler(Math.random() * 10, Math.random() * 10, Math.random() * 10)
-	}
-
-	type Body = {
-		id: string
-		mounted: number
-		position: Vector3
-		rotation: Euler
-	}
-
-	let bodies: Body[] = []
-
-	let lastBodyMounted: number = 0
-	let bodyEveryMilliseconds = 50
-	let longevityMilliseconds = 8000
-
-	useFrame(() => {
-		if (lastBodyMounted + bodyEveryMilliseconds < Date.now()) {
-			const body: Body = {
-				id: getId(),
-				mounted: Date.now(),
-				position: getRandomPosition(),
-				rotation: getRandomRotation()
-			}
-			bodies.unshift(body)
-			lastBodyMounted = Date.now()
-			bodies = bodies
-		}
-		const deleteIds: string[] = []
-		bodies.forEach((body) => {
-			if (body.mounted + longevityMilliseconds < Date.now()) {
-				deleteIds.push(body.id)
-			}
-		})
-
-		if (deleteIds.length) {
-			deleteIds.forEach((id) => {
-				const index = bodies.findIndex((body) => body.id === id)
-				if (index !== -1) bodies.splice(index, 1)
-			})
-			bodies = bodies
-		}
-	})
 </script>
 
 <PerspectiveCamera position={{ y: 10, z: 10, x: 10 }}>
@@ -83,6 +23,4 @@
 
 <Debug />
 
-{#each bodies as body (body.id)}
-	<TestBody position={body.position} rotation={body.rotation} />
-{/each}
+<Emitter />
